Move all documentation content into the docs dir
In order to keep things as clean as is possible, this adjusts the
documentation and configuration for building it so that all of the
required files are wholly contained only within the docs folder,
instead of having the config outside of it.

This makes for a more natural documentation setup and a cleaner way to
elide the documentation.

In order to set this up, the build needs to change because Cloudflare
uses a chroot jail (seemingly) to isolate the build to the specified
working directory, but it does this after it has checked to see if there
is a requirements file.

So, this also includes a simple driver script which can ensure that the
build carries out correctly in the docs folder while not setting that
as the root.

As an upside to this, we can now unshallow the git clone so that each
documentation file gets a unique date based on its creation, which is
not possible with the HEAD-only commit that it usually does to save
time.