Implement the start of a template expansion script
Our next goal is to have a script that's packed into the repository that
can either copy our content to a new package folder, copy the content
from some arbitrary folder to a newly created package folder, or clone
a repository into a newly created package folder.

Either way, once the source files are known, the newly created package
will have its contents expanded out for various placeholders, getting it
ready to be a new Sublime package.

This version of the script has all of the structure required to
determine where the source and destination should be, and what files
should be involved.

As of this commit no actual template expansions happen, although we do
have the logic in place to copy files and ensure that if we're doing an
in place expansion that the files are given temporary names and then
moved into place.

The outstanding portion of this is:
  - Detecting when a file is not text, so we can just copy it
  - Doing something like shutil.copy(), but modifying the content as it
    is copied.