Include actual template expansion logic
This includes the logic required to actually handle text files and
expand variables from within them out. The transform of such files is
done by loading the content and altering it as needed before writing it
to the destination; we ensure that the properties of the source are
propagated to the destination.

There is currently a small list of predefined variables for use here,
which can be augmented later.