Do slight cleanup of template code
This cleans up a couple of small issues relating to a type annotation
that we forgot to change since we returned a filter and not a list, and
clean up some unused arguments.

We also fixed up the config file, which should have had quotes around
the string variables that now have templates so that the site builder
does not yack.